organized new classes and tests into desire hierarchy
