+++ Calculator.js
@@ -1,6 +1,6 @@
 const MathOperations = require('./MathOperations.js');
-
 class Calculator{
+
     add(a = null, b = null) {
         if (Array.isArray(a)) {
             return this.result = MathOperations.sum(a);
@@ -33,21 +33,6 @@ class Calculator{
     sumList(array){
         return this.result = MathOperations.sumList(array)
     }
-
-    mean(array){
-        return this.result = MathOperations.mean(array)
-    }
-
-    mode(array){
-        return this.result = MathOperations.mode(array)
-    }
-
-    standardDeviation(array){
-        return this.result = MathOperations.standardDeviation(array)
-    }
-
-
-
 
 }
