+++ MathOperations.js
@@ -4,9 +4,6 @@ const division = require('./division');
 const subtraction = require('./subtraction');
 const squaring = require('./squaring');
 const squareRooting = require('./squareRooting');
-const meanStats = require('./meanStats');
-const modeStats = require('./modeStats');
-const standardDeviation = require('./standardDeviation');
 
 class MathOperations {
     static sum(a = null, b = null) {
@@ -42,17 +39,6 @@ class MathOperations {
 
     }
 
-    static mean(myarray = []){
-        return meanStats.mean(myarray)
-    }
-
-    static mode(myarray = []){
-        return modeStats.mode(myarray)
-    }
-
-    static standardDeviation(myarray = []){
-        return standardDeviation.standardDeviation(myarray)
-    }
 }
 
 module.exports = MathOperations;
